Added disagreement revisions to Identify.
Also refactored a bit to avoid repetition. Not sure it's really better as it
adds some artificial attributes to identifications which might be a bit
mysterious. Time will tell.
